Implementation of Display Object MANAGER function
Moved components out of the main project file into a display object
manager function which, itself, currently lives in the main project
file.  Vectors/bins of objects are hard coded instantiated and in the
main while loop, update and update touch is run. THen, pointers to these
bins are passed into the manager which does all the work and spits
results back into the bins.